Voir une ressource authentifié ou non

+++ app/Http/Controllers/API/RessourceController.php
@@ -13,6 +13,7 @@ use App\Repositories\LienRessourceRepository;
 use Illuminate\Http\Request;
 use App\Repositories\RessourceRepository;
 use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class RessourceController extends Controller
 {
@@ -26,7 +27,9 @@ class RessourceController extends Controller
         $this->ressourceRepository = $ressourceRepository;
         $this->lienRessourceRepository = $lienRessourceRepository;
         $this->favorisRepository = $favorisRepository;
+
     }
+
 
 
     /**
@@ -36,38 +39,41 @@ class RessourceController extends Controller
      */
     public function index(Request $request)
     {
-        
+
         $validated = $request->validate([
             'type' => 'sometimes|int',
             'categorie' => 'sometimes|int',
         ]);
-        
+
         $lesRessources = $this->ressourceRepository->allPublic();
-        
-        
-        $idType=null;
-        $idCategorie=null;
 
-        try{
+
+        $idType = null;
+        $idCategorie = null;
+
+        try {
             $idType = $validated['type'];
-        }catch(Exception $e){}
-        try{
+        } catch (Exception $e) {
+        }
+        try {
             $idCategorie = $validated['categorie'];
-        }catch(Exception $e){}
-        
+        } catch (Exception $e) {
+        }
 
-        if($idType!=null){
+
+        if ($idType != null) {
             $lesRessources = $this->ressourceRepository->AllPublicByType($idType);
         }
-        if($idCategorie!=null){
-            if($idType!=null){
+        if ($idCategorie != null) {
+            if ($idType != null) {
                 $lesRessources = $this->ressourceRepository->AllPublicByTypeAndCateg($idType, $idCategorie);
-            }else{
+            } else {
                 $lesRessources = $this->ressourceRepository->AllPublicByCategorie($idCategorie);
             }
         }
-         
+
         
+
         return response()->json($lesRessources);
     }
 
@@ -76,11 +82,11 @@ class RessourceController extends Controller
     public function indexUtilisateur(Request $request)
     {
         $authID = auth()->user()->id;
-        $result = $this->ressourceRepository->allPartages($authID);
-        if ($request->favori) {
+        $result = $this->ressourceRepository->allPrivees($authID);
+        if ($request->favoris) {
             $result = $result->merge($this->favorisRepository->whereFavoris($authID));
         }
-        if ($request->partagee) {
+        if ($request->partagees) {
             $result = $result->merge($this->lienRessourceRepository->findRessourceByIdUser($authID));
         }
         if ($request->mise_de_cote) {
@@ -100,8 +106,9 @@ class RessourceController extends Controller
     public function store(StoreUpdateRessourceRequest $request)
     {
         $validated = $request->validated();
+        $result = array_merge($validated, ['IdUtilisateur_createur' => auth()->user()->id]);
 
-        Ressources::create($validated);
+        Ressources::create($result);
     }
 
     /**
@@ -110,13 +117,40 @@ class RessourceController extends Controller
      * @param  \App\Models\Ressources  $ressource
      * @return \Illuminate\Http\Response
      */
-    public function show(Ressources $ressource)
+    public function show(Ressources $ressource, Request $request)
     {
-        $laRessource = [
-            'ressource' => $this->ressourceRepository->one($ressource)
-        ];
-        return response()->json($laRessource);
+
+        $autorisation=false;
+        $result =["message" => "unauthorized"];
+        if($ressource->IdEtat==4){
+            $autorisation=true;
+        }
+        elseif(auth('api')->user()!=null){
+            $authID=auth('api')->user()->id;
+            $createurRessource = $ressource->Utilisateur()->get();
+            if($createurRessource[0]['id'] == $authID){
+                $autorisation=true;
+            }else{
+                $lesRessourcesPartages = $this->lienRessourceRepository->findRessourceByIdUser($authID);
+                foreach($lesRessourcesPartages as $uneRessource){
+                    if($uneRessource['id'] == $ressource->id){
+                        $autorisation=true;
+                    }
+                }
+            }
+        }
+
+        if($autorisation){
+            $result = $this->ressourceRepository->one($ressource);
+        }
+
+        // return response()->json($result);
+        return response()->json($result);
+        // return auth('api')->user();
+        ;
     }
+
+
 
     /**
      * Update the specified resource in storage.
@@ -127,10 +161,13 @@ class RessourceController extends Controller
      */
     public function update(StoreUpdateRessourceRequest $request, $id)
     {
-        $ressource = Ressources::find($id);
+        $ressourceAModif = Ressources::find($id);
+        $ressource = $this->ressourceRepository->findById($id);
         $validated = $request->validated();
-
-        $ressource->update($validated);
+        if ($this->ressourceRepository->findCreateur($ressource['id']) == auth()->user()->id) {
+            //Si elle est privé, il faut supprimer les liens avec
+            $ressourceAModif->update($validated);
+        }
     }
 
     /**
@@ -142,14 +179,16 @@ class RessourceController extends Controller
     public function destroy($id)
     {
         $ressource = $this->ressourceRepository->findById($id);
-        $lesLiensRessource = $this->lienRessourceRepository->findByIdRessource($id);
+        if ($this->ressourceRepository->findCreateur($ressource['id']) == auth()->user()->id) {
+            $lesLiensRessource = $this->lienRessourceRepository->findByIdRessource($id);
 
-        foreach ($lesLiensRessource as $unLienRessource) {
-            Jointure_ress_utilisateur::destroy($unLienRessource['id']);
+            foreach ($lesLiensRessource as $unLienRessource) {
+                Jointure_ress_utilisateur::destroy($unLienRessource['id']);
+            }
+            foreach ($ressource['commentaires'] as $unCommentaire) {
+                Commentaire::destroy($unCommentaire['id']);
+            }
+            Ressources::destroy($id);
         }
-        foreach ($ressource['commentaires'] as $unCommentaire) {
-            Commentaire::destroy($unCommentaire['id']);
-        }
-        Ressources::destroy($id);
     }
 }
